refactor: Professionalize project structure and add config files
This commit introduces several changes to improve the project's organization and make it align with common Python project standards:

1.  **Directory Restructuring:**
    *   All core Python source files (`main.py`, `snake.py`, `food.py`, `gameboard.py`, `ui_manager.py`) have been moved into a `src/` directory.
    *   `src/__init__.py` was added to make `src` a Python package.

2.  **Import Updates:**
    *   All import statements within the `src` directory and the `tests` directory have been updated to reflect the new structure. Source files now use relative imports (e.g., `from .food import Food`), and test files import from the `src` package (e.g., `from src.snake import Snake`).
    *   The path to `config.json` in `src/main.py` was adjusted to correctly point to the project root.

3.  **Added `requirements.txt`:**
    *   A `requirements.txt` file has been added to the project root, specifying `pygame>=2.0.0` as a dependency. This facilitates easier setup of your project environment.

4.  **Added `.gitignore`:**
    *   A comprehensive `.gitignore` file has been added to the project root. This file includes common patterns for Python-generated files, virtual environments, OS-specific files, and IDE/editor configuration files to keep the repository clean.

These changes provide a more professional and maintainable structure for the project. Running the game is now typically done via `python -m src.main` and tests via `python -m unittest discover -s tests` from the project root.
